public shape loader functions
